refactor CustomTable -> Table
