--- sp1-custom-table/src/app/shared/components/custom-table/custom-table.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CustomTableComponent } from './custom-table.component';
-
-describe('CustomTableComponent', () => {
-  let component: CustomTableComponent;
-  let fixture: ComponentFixture<CustomTableComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [CustomTableComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(CustomTableComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
