+++ sp1-custom-table/src/app/shared/components/table/child-components/modify-columns/modify-columns.component.scss
