+++ sp1-custom-table/src/app/shared/components/table/models/auto-refresh.model.ts
