+++ sp1-custom-table/src/app/shared/components/table/models/sort.model.ts
