+++ sp1-custom-table/src/app/shared/components/table/models/table.model.ts
