+++ sp1-custom-table/src/app/shared/components/table/services/row-selection.service.ts
