+++ sp1-custom-table/src/app/shared/components/table/table.component.scss
