+++ sp1-custom-table/src/app/shared/components/table/table.module.ts
@@ -48,4 +48,4 @@ const material = [
     PathValuePipe,
   ],
 })
-export class CustomTableModule { }
+export class TableModule { }
